refactor: extract SSL-mode migration into a path-based shared module
Moves stale_postgres_ssl_mode_replacement, migrate_connection_ssl_mode_in_place,
and their tests out of commands.rs into a new connection_migrations module.
The orchestration function (migrate_postgres_ssl_mode_spelling) is split into
a path-based core (migrate_postgres_ssl_mode_spelling_at_path, no AppHandle
dependency) plus a thin AppHandle wrapper left in commands.rs that just
resolves the path and invalidates the connection cache.

Per .rules/rust.md rule 1 (extract pure helpers into dedicated sibling
modules) and to avoid a one-directional dependency between commands.rs
(6300+ lines) and mcp/mod.rs — this is prep for wiring the migration into the
--mcp server process (issue #639), which has no AppHandle. That wiring is a
separate follow-up commit; this one is a pure extraction with no behavior
change, except the return type improves from Result<()> to Result<bool> so
callers can tell "nothing needed migrating" apart from "migrated and saved."
